lab3/backend/requirements.txt
pika
psycopg2-binary
